@@ -1,100 +1,86 @@
 {
-  "name": "signeoappfront",
-  "author": "Signeo",
-  "description": "Signeo Client App",
-  "version": "0.1.0",
+  "name": "my-electron",
   "private": true,
-  "main": "public/electron.js",
-  "homepage": ".",
-  "dependencies": {
-    "@babel/core": "^7.0.0-0",
-    "@babel/plugin-syntax-flow": "^7.14.5",
-    "@babel/plugin-transform-react-jsx": "^7.14.9",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "@mui/icons-material": "^5.15.20",
-    "@mui/material": "^5.15.20",
-    "@mui/styles": "^5.15.20",
-    "@testing-library/dom": "^10.2.0",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^13.0.0",
-    "@testing-library/user-event": "^13.2.1",
-    "dmg-builder": "24.13.3",
-    "electron-is-dev": "^3.0.1",
-    "postcss": "^8.1.0",
-    "prop-types": "^15.8.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-router-dom": "^6.24.0",
-    "react-scripts": "^5.0.1",
-    "typescript": "^5.5.2",
-    "web-vitals": "^2.1.0"
-  },
+  "version": "0.0.0",
+  "type": "module",
+  "main": "dist/electron/main.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "electron-dev": "concurrently \"cross-env BROWSER=none yarn start\" \"wait-on http://localhost:3000 && cross-env NODE_ENV=development electron .\" \"wait-on http://localhost:3000 && cross-env NODE_ENV=development node server.js\"",
-    "electron-pack": "concurrently \"cross-env NODE_ENV=production yarn build\" \"wait-on build/index.html && cross-env NODE_ENV=production electron .\"",
-    "dist": "cross-env NODE_ENV=production yarn build && electron-builder",
-    "backend": "node server.js"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@babel/plugin-transform-private-property-in-object": "^7.24.7",
-    "@eslint/compat": "^1.1.0",
-    "@eslint/js": "^9.6.0",
-    "autoprefixer": "^10.4.19",
-    "concurrently": "^8.2.2",
-    "cross-env": "^7.0.3",
-    "dotenv": "^16.4.5",
-    "electron": "^31.1.0",
-    "electron-builder": "^24.13.3",
-    "electron-builder-squirrel-windows": "24.13.3",
-    "eslint": "^9.6.0",
-    "eslint-config-react-app": "^7.0.1",
-    "eslint-plugin-jest": "^28.6.0",
-    "eslint-plugin-react": "^7.34.3",
-    "globals": "^15.8.0",
-    "postcss-loader": "^8.1.1",
-    "prettier": "^3.3.2",
-    "tailwindcss": "^3.4.4",
-    "wait-on": "^7.2.0"
+    "dev": "vite",
+    "build": "vite build && electron-builder",
+    "lint": "eslint . --ext .ts,.tsx,.js,.jsx",
+    "preview": "vite preview"
   },
   "build": {
-    "appId": "com.signeo.signeoappfront",
-    "productName": "Signeo Client",
+    "appId": "com.example.myapp",
+    "productName": "My_Client",
+    "asar": true,
     "files": [
-      "build/**/*",
-      "public/**/*",
-      "package.json"
+      "dist/renderer/**",
+      "dist/electron/**"
     ],
     "directories": {
-      "buildResources": "assets"
+      "buildResources": "assets",
+      "output": "dist/release"
     },
     "win": {
-      "target": "nsis"
+      "icon": "src/assets/icon/test.ico",
+      "target": "portable"
     },
     "mac": {
+      "icon": "src/assets/icon/test.icns",
       "target": "dmg"
     },
     "linux": {
+      "icon": "src/assets/icon/test.png",
       "target": "AppImage"
-    }
+    },
+    "portable": {
+      "artifactName": "${productName}-${version}-portable.exe"
+    },
+    "extraResources": [
+      {
+        "from": "src/assets/icon/test.ico",
+        "to": "icon.ico"
+      }
+    ]
   },
-  "repository": "https://github.com/SIGNEO/SigneoAppFront.git",
-  "license": "MIT"
+  "dependencies": {
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^7.1.2",
+    "@mui/lab": "7.0.0-beta.14",
+    "@mui/material": "^7.1.2",
+    "@mui/styles": "^6.4.8",
+    "axios": "^1.10.0",
+    "electron-store": "^10.1.0",
+    "prop-types": "^15.8.1",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.2"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.29.0",
+    "@tailwindcss/vite": "^4.1.10",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "@vitejs/plugin-react": "^4.6.0",
+    "electron": "^36.5.0",
+    "electron-builder": "^26.0.12",
+    "eslint": "^9.29.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.20",
+    "globals": "^16.2.0",
+    "tailwindcss": "^4.1.10",
+    "typescript": "~5.8.3",
+    "typescript-eslint": "^8.35.0",
+    "vite": "^6.3.5",
+    "vite-plugin-electron": "^0.29.0",
+    "vite-plugin-electron-renderer": "^0.14.6"
+  },
+  "packageManager": "pnpm@10.12.2",
+  "license": "MIT",
+  "author": {
+    "name": "Ignacio"
+  },
+  "description": "My Electron App"
 }
